starting basic structure of the properties databases
