+++ oscps-lib/src/properties/pure_species_properties.rs
@@ -1,0 +1,8 @@
+///# Pure Species Properties
+/// This sub-module will contain the information regarding the cheracteristic properties of pure
+/// species. This would include the molar mass (molecular weight), accentric factor, critical
+/// temperature, critical pressure, critical compressibility, critical molar volume, and normal
+/// boiling point.
+
+
+use crate::properties::*;
